Apply nullboot purge/reinstall and FDE kernel image only for CVM

+++ vhdbuilder/packer/pre-install-dependencies.sh
@@ -141,11 +141,21 @@ if [[ ${UBUNTU_RELEASE//./} -ge 2204 && "${ENABLE_FIPS,,}" != "true" ]]; then
       "linux-cloud-tools-5.15.0-1092-azure"
     )
     echo "Ubuntu 22.04 x86_64 detected, installing pinned kernel version 5.15.0-1092"
+  elif grep -q "cvm" <<< "$FEATURE_FLAGS"; then
+    KERNEL_IMAGE="linux-image-azure-fde-lts-${UBUNTU_RELEASE}"
+    KERNEL_PACKAGES=(
+      "linux-image-azure-fde-lts-${UBUNTU_RELEASE}"
+      "linux-tools-azure-lts-${UBUNTU_RELEASE}"
+      "linux-cloud-tools-azure-lts-${UBUNTU_RELEASE}"
+      "linux-headers-azure-lts-${UBUNTU_RELEASE}"
+      "linux-modules-extra-azure-lts-${UBUNTU_RELEASE}"
+    )
+    echo "Installing fde LTS kernel for CVM Ubuntu ${UBUNTU_RELEASE}"
   else
     # Use LTS kernel for other versions  
     KERNEL_IMAGE="linux-image-azure-lts-${UBUNTU_RELEASE}"
     KERNEL_PACKAGES=(
-      "linux-image-azure-fde-lts-${UBUNTU_RELEASE}"
+      "linux-image-azure-lts-${UBUNTU_RELEASE}"
       "linux-tools-azure-lts-${UBUNTU_RELEASE}"
       "linux-cloud-tools-azure-lts-${UBUNTU_RELEASE}"
       "linux-headers-azure-lts-${UBUNTU_RELEASE}"
@@ -160,9 +170,11 @@ if [[ ${UBUNTU_RELEASE//./} -ge 2204 && "${ENABLE_FIPS,,}" != "true" ]]; then
   if apt-cache show "$KERNEL_IMAGE" &>/dev/null; then
     echo "Kernel packages are available, proceeding with purging current kernel and installing new kernel..."
 
-    # Purge nullboot package
-    wait_for_apt_locks
-    DEBIAN_FRONTEND=noninteractive apt-get remove --purge -y --allow-remove-essential nullboot
+    # Purge nullboot package only for cvm
+    if grep -q "cvm" <<< "$FEATURE_FLAGS"; then
+      wait_for_apt_locks
+      DEBIAN_FRONTEND=noninteractive apt-get remove --purge -y --allow-remove-essential nullboot
+    fi
 
     # Purge all current kernels and dependencies
     wait_for_apt_locks
@@ -174,9 +186,11 @@ if [[ ${UBUNTU_RELEASE//./} -ge 2204 && "${ENABLE_FIPS,,}" != "true" ]]; then
     DEBIAN_FRONTEND=noninteractive apt-get install -y "${KERNEL_PACKAGES[@]}"
     echo "After installing new kernel, here is a list of kernels/headers installed:"; dpkg -l 'linux-*azure*'
 
-    # Reinstall nullboot package
-    wait_for_apt_locks
-    DEBIAN_FRONTEND=noninteractive apt-get install -y nullboot
+    # Reinstall nullboot package only for cvm
+    if grep -q "cvm" <<< "$FEATURE_FLAGS"; then
+      wait_for_apt_locks
+      DEBIAN_FRONTEND=noninteractive apt-get install -y nullboot
+    fi
   else
     echo "Kernel packages for Ubuntu ${UBUNTU_RELEASE} are not available. Skipping purging and subsequent installation."
   fi
